Generate the Makefile. It will now be placed in the bindir and will be populated with all the known information. Beware: one still has to manually set the CUDA lib and path as they are not available after configure (unlike the include which is).
Conflicts:
	opal/datatype/cuda/Makefile

This file was certainly not supposed to be here. There is NO valid
reason to have a copy of a locally generated file in the source.

Add the capability to install the generated library and other
minor cleanups.

Open the datatype CUDA library from a default install location.
Various other minor cleanups.